Adding application universe for Users

+++ study-go-kotlin-spring-boot/src/main/kotlin/com/eduardo/main/repository/UserRepository.kt
@@ -1,8 +1,8 @@
 package com.eduardo.main.repository
 
-import com.eduardo.main.model.User
+import com.eduardo.main.model.database.User
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
 
 @Repository
-interface UserRepository : JpaRepository<User, Int>
+interface UserRepository : JpaRepository<User, Long>
